debug regime value clipping

+++ include/universal/posit/regime.hpp
@@ -45,7 +45,7 @@ public:
 		long double scale;
 		int e2 = (1 << es) * _k;
 		if (e2 < -63 || e2 > 63) {
-			scale = std::pow(2.0, e2);  // TODO: needs a native posit implementation
+			scale = std::pow(2.0l, (long double)e2);  // TODO: needs a native posit implementation
 		}
 		else {
 			if (e2 >= 0) {
